// src/app/@core/enums/xlsx-template.enum.ts
export enum XlsxTemplateEnum {
    BasicReport = './assets/excel-templates/basic.xlsx',
    AuditReport = './assets/excel-templates/audit-report.xlsx',
    LoadPointReport = './assets/excel-templates/loadpoint-report.xlsx',
    LoadPointsSummaryReport = './assets/excel-templates/loadpoints-summary-report.xlsx',
    GeneratingUnitsSummaryReport = './assets/excel-templates/generating-units-summary-report.xlsx',
    GeneratingUnitReport = './assets/excel-templates/generating-unit-report.xlsx',
    PowerStationsReport = './assets/excel-templates/power-stations-report.xlsx',
    PowerStationsSummaryReport = './assets/excel-templates/power-stations-summary-report.xlsx',
}